# apps/api/openapi/order-tracking.yaml
openapi: 3.0.3
info:
  title: Customer Order Read and Tracking
  version: 1.0.0
  description: Customer-safe order projection. Identifiers and phone numbers never grant access.
paths:
  /ordering/orders/{idOrOrderNumber}:
    get:
      summary: Read an order owned by the authenticated customer
      security:
        - customerSession: []
      parameters:
        - $ref: '#/components/parameters/orderLookup'
      responses:
        '200':
          $ref: '#/components/responses/tracking'
        '401':
          description: Missing or invalid customer session
        '404':
          description: Order not found or not owned by this customer
  /ordering/orders/{idOrOrderNumber}/tracking:
    get:
      summary: Track an order owned by the authenticated customer
      security:
        - customerSession: []
      parameters:
        - $ref: '#/components/parameters/orderLookup'
      responses:
        '200':
          $ref: '#/components/responses/tracking'
        '401':
          description: Missing or invalid customer session
        '404':
          description: Order not found or not owned by this customer
  /ordering/guest/orders/{orderNumber}/tracking:
    get:
      summary: Track one guest order with its opaque order access token
      description: The token is returned at guest order submission and on an authorized idempotent replay. Keep it out of URLs and logs. It expires after 90 days and may be revoked.
      security:
        - guestOrderToken: []
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
        - name: x-order-access-token
          in: header
          required: true
          schema:
            type: string
            minLength: 43
            maxLength: 43
      responses:
        '200':
          $ref: '#/components/responses/tracking'
        '401':
          description: Missing or malformed token
        '404':
          description: Order/token pair not found, expired, or revoked
components:
  securitySchemes:
    customerSession:
      type: http
      scheme: bearer
      description: Authenticated customer session.
    guestOrderToken:
      type: apiKey
      in: header
      name: x-order-access-token
  parameters:
    orderLookup:
      name: idOrOrderNumber
      in: path
      required: true
      schema:
        type: string
  responses:
    tracking:
      description: Customer-safe order and tracking projection
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CustomerOrderTracking'
  schemas:
    CustomerOrderTracking:
      type: object
      required: [orderNumber, createdAt, status, orderReceived, availabilityConfirmed, customerName, deliveryAddress, amounts, items, timeline, delivery, actionRequired, resolutions]
      properties:
        orderNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/CustomerStage'
        orderReceived:
          type: boolean
        availabilityConfirmed:
          type: boolean
          description: True only when the order or fulfillment facts support confirmation.
        customerName:
          type: string
        deliveryAddress:
          type: object
          properties:
            governorate: { type: string, nullable: true }
            area: { type: string, nullable: true }
            address: { type: string }
            latitude: { type: number, nullable: true }
            longitude: { type: number, nullable: true }
            locationSource: { type: string, nullable: true }
        amounts:
          type: object
          required: [originalCodAmount, currentCodAmount, subtotal, shippingFee, currency]
          properties:
            originalCodAmount: { type: number, description: Immutable order snapshot. }
            currentCodAmount: { type: number, description: Current collection expectation when available; otherwise original COD. }
            subtotal: { type: number }
            shippingFee: { type: number }
            currency: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/CustomerOrderItem'
        timeline:
          type: array
          description: Persisted order, shipment, and cancellation events only.
          items:
            type: object
            required: [status, occurredAt]
            properties:
              status:
                $ref: '#/components/schemas/CustomerStage'
              occurredAt:
                type: string
                format: date-time
        delivery:
          type: object
          required: [certainty, estimatedDate, deliveredAt, source, nextUpdateBy, promise]
          properties:
            certainty:
              type: string
              enum: [unknown, estimated, confirmed]
            estimatedDate: { type: string, format: date-time, nullable: true }
            deliveredAt: { type: string, format: date-time, nullable: true }
            source: { type: string, nullable: true }
            nextUpdateBy: { type: string, format: date-time, nullable: true }
            promise: { type: string, nullable: true }
        actionRequired:
          nullable: true
          description: Null until a canonical customer-action workflow exists.
          allOf:
            - $ref: '#/components/schemas/CustomerAction'
        resolutions:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResolution'
    CustomerStage:
      type: string
      enum: [RECEIVED, CONFIRMING_PRODUCTS, CONFIRMED, PREPARING, OUT_FOR_DELIVERY, DELIVERED, ACTION_REQUIRED, CANCELLATION_REQUESTED, CANCELLED, DELIVERY_ISSUE]
    CustomerOrderItem:
      type: object
      required: [id, skuId, productName, skuCode, variantName, quantity, unitPrice, discountAmount, lineTotal, currency, status]
      properties:
        id: { type: string, format: uuid }
        skuId: { type: string, format: uuid }
        productName: { type: string, nullable: true }
        skuCode: { type: string, nullable: true }
        variantName: { type: string, nullable: true }
        quantity: { type: integer }
        unitPrice: { type: number }
        discountAmount: { type: number }
        lineTotal: { type: number }
        currency: { type: string }
        status:
          type: string
          enum: [checking, confirmed, action-required, resolved]
    CustomerAction:
      type: object
      required: [category, affectedItemIds, allowedActions]
      properties:
        category: { type: string }
        affectedItemIds:
          type: array
          items: { type: string, format: uuid }
        allowedActions:
          type: array
          items: { type: string }
    CustomerResolution:
      type: object
      required: [category, status, affectedItemIds, allowedActions]
      properties:
        category:
          type: string
          enum: [cancellation, return, refund, replacement, compensation]
        status:
          type: string
          enum: [requested, in_progress, resolved, rejected]
        affectedItemIds:
          type: array
          items: { type: string, format: uuid }
        allowedActions:
          type: array
          items: { type: string }